[518755] Interrupt merge on errors in the comparison
If there is an error diagnostic in the comparison model, we should stop
the merge operation, as there are only errors in the comparison model if
something severely went wrong during the comparison. Thus, the detected
differences will probably make no sense anyway and we risk breaking the
model if we don't stop the comparison and report an error, because if we
just continue, EGit will report a successful merge even though we didn't
perform the merge correctly.

Note that we still don't (and don't want to) stop if there are
diagnostic errors in one or more of the compared models (in contrast to
the comparison). Errors in the models most likely don't stop us from
doing a decent merge.

Bug: 518755
Change-Id: I209505dec67daed825f4ba63f40b1f10a79ae36c
Signed-off-by: Philip Langer <planger@eclipsesource.com>