Register the diagram CMV for any diff of diagram elements
Several changes in diagrams, such as style changes in Papyrus diagrams,
won't be opened in the DiagramContentMergeViewer. Before this change,
the DiagramContentMergeViewer was only used for DiagramDiffs. However,
it doesn't make sense to produce a DiagramDiff for _any_ diff that is
applied in a diagram, as it would unnecessarily multiply the number of
diffs. Instead, it would be better to just open the Diagram CMV for any
diff that has been applied to a diagram element and make the accessors
more flexible so that they are able to retrieve the view for any diff
in a diagram.

Change-Id: Ib2431819a22ece745a2e18a8462095c483c44c1f
Signed-off-by: Philip Langer <planger@eclipsesource.com>