CMV switcher menu has duplicates viewing a string attribute diff
This happens because CompareUIPlugin#guessType guesses that it's a TXT
type and shows the default text viewer for that. This is confusing and
useless.

We seem to be only able to avoid that by inspecting the call stack. We
should avoid returning contents also for getContentType because we don't
ever want a computed content type to be used to determine which content
viewer is available.

Change-Id: I5eb747b32d8f22e971a06639777bb8ce7067d70f
Signed-off-by: Philip Langer <planger@eclipsesource.com>